# swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/ordenadores/contrato/{id}": {
      "get": {
        "operationId": "OrdenadoresController_consultarInfoOrdenador",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del contrato general",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna detalle de un supervisor."
          },
          "400": {
            "description": "Error en la solicitud."
          }
        },
        "summary": "Retorna información de Supervisor por ID de contrato.",
        "tags": [
          "Ordenadores"
        ]
      }
    },
    "/supervisores/contrato/{id}": {
      "get": {
        "operationId": "SupervisoresController_consultarInfoSupervisor",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del contrato general",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retorna detalle de un supervisor."
          },
          "400": {
            "description": "Error en la solicitud."
          }
        },
        "summary": "Retorna información de Supervisor por ID de contrato.",
        "tags": [
          "Supervisores"
        ]
      }
    }
  },
  "info": {
    "title": "Gestion Contractual API MID",
    "description": "API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
...
}